# api/docs/openapi.yml
...
openapi: 3.1.0
info:
  title: OpenTalk RoomServer API
  description: Specifies the endpoints and structure of the OpenTalk RoomServer Web API
  contact:
    name: OpenTalk Team
    email: mail@opentalk.eu
  license:
    name: EUPL-1.2
    identifier: EUPL-1.2
  version: 0.0.6-dev
servers:
  - url: /v1
paths:
  "/rooms/{room_id}":
    put:
      tags:
        - "v1::rooms"
      summary: Creates a new room instance with the specified parameters if no room with the provided id exists.
      description: "If a room with the provided room ID already exists, the rooms idle timeout is refreshed."
      operationId: put_room
      parameters:
        - name: room_id
          in: path
          description: The UUID that identifies the room
          required: true
          schema:
            $ref: "#/components/schemas/RoomId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomParameters"
        required: true
      responses:
        "201":
          description: Successfully created a new room
        "204":
          description: The room did exist before and the parameter were updated if necessary
        "400":
          description: The provided API token could not be parsed
        "401":
          description: The provided API token is invalid
        "422":
          description: Failed to parse request body
        "500":
          description: An internal server error occurred
      security:
        - API-Token: []
  "/rooms/{room_id}/token":
    post:
      tags:
        - "v1::rooms"
      summary: Creates a new signaling token for the specified user and room
      description: |-
        The signaling token can be used to establish a websocket connection with the roomserver through the
        `/signaling/<token>` endpoint. The token has a limited lifetime (30 seconds by default) and can only be used once.

        Calling this endpoint will start a new room task or refresh existing ones. To get a token for an unknown room, the
        request body has to contain the `room_parameters` field (See [`TokenRequestBody`]). If the room is already running,
        any provided `room_parameters` will be ignored.
      operationId: request_token
      parameters:
        - name: room_id
          in: path
          description: The UUID that identifies the room
          required: true
          schema:
            $ref: "#/components/schemas/RoomId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequestBody"
        required: true
      responses:
        "200":
          description: "The response body contains either the signaling token, or a prompt to include the associated room parameters in the request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: The provided API token could not be parsed
        "401":
          description: The provided API token is invalid
        "422":
          description: Failed to parse request body
        "500":
          description: An internal server error occurred
      security:
        - API-Token: []
  "/signaling/{token}":
    get:
      tags:
        - "v1::signaling"
      summary: Opens a new signaling websocket connection.
      operationId: open_signaling_socket
      parameters:
        - name: token
          in: path
          description: The UUID token that verifies the user
          required: true
          schema:
            $ref: "#/components/schemas/RoomId"
      responses:
        "101":
          description: Successfully upgraded connection to WebSocket
        "404":
          description: The requested room does not exist
        "500":
          description: An internal server error occurred
      security: []
components:
  schemas:
    CallInId:
      $ref: "#/components/schemas/NumericId"
      description: The id of a call-in participation
    CallInInfo:
      type: object
      description: Information needed to participate in a call-in connection.
      required:
        - tel
        - id
        - password
      properties:
        id:
          $ref: "#/components/schemas/CallInId"
          description: SIP ID which must transmitted via DTMF (number field on the phone) to identify this room
        password:
          $ref: "#/components/schemas/CallInPassword"
          description: |-
            SIP password which must be transmitted via DTMF (number field on the phone) after entering the `sip_id`
            to enter the room
        tel:
          type: string
          description: SIP Call-In phone number which must be used to reach the room
      example:
        id: "1234567890"
        password: "0987654321"
        tel: +555-123-456-789
    CallInPassword:
      $ref: "#/components/schemas/NumericId"
      description: The password for authenticating call-in participation
...
    ClientParameters:
      allOf:
        - $ref: "#/components/schemas/ClientKind"
          description: Information about a registered user
        - type: object
          required:
            - device_secret
            - role
          properties:
            device_secret:
              $ref: "#/components/schemas/DeviceSecret"
              description: "Identifier to maintain participant state across reconnections, is generated by the client"
            role:
              $ref: "#/components/schemas/Role"
              description: The role that the client assumes in the meeting.
      description: Client context required for a signaling connection
      example:
        device_secret: v3rys3cr3tD3v1ce5tr1ng
        kind: registered
        profile:
          avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
          display_name: Alice Adams
          email: alice@example.com
          firstname: Alice
          id: 00000000-0000-0000-0000-0000000a11c3
          lastname: Adams
          timezone: Europe/Berlin
          title: ""
        role: moderator
...
    EventContext:
      type: object
      required:
        - id
        - title
        - description
        - is_adhoc
      properties:
        description:
          $ref: "#/components/schemas/EventDescription"
        ends_at:
          type:
            - string
            - "null"
          format: date-time
        id:
          $ref: "#/components/schemas/EventId"
        is_adhoc:
          type: boolean
        shared_folder:
          $ref: "#/components/schemas/SharedFolder"
        starts_at:
          type:
            - string
            - "null"
          format: date-time
        title:
          $ref: "#/components/schemas/EventTitle"
      example:
        description: The Weekly Team Event
        ends_at: "1970-01-01T01:00:00Z"
        id: 00000000-0000-0000-0000-004433221100
        is_adhoc: false
        shared_folder:
          read:
            password: v3rys3cr3t
            url: "https://cloud.example.com/shares/abc123"
        starts_at: "1970-01-01T00:00:00Z"
        title: Team Event
    EventDescription:
      type: string
      description: The description of an event
      examples:
        - The Weekly Team Event
      maxLength: 4096
    EventId:
      type: string
      format: uuid
      description: The identifier of an event
      example: 00000000-0000-0000-0000-004433221100
    EventTitle:
      type: string
      description: The title of an event
      examples:
        - Team Event
      maxLength: 255
...
    InviteCode:
      type: string
      format: uuid
      description: An invite code
      example: 00000000-0000-0000-0000-0000deadbeef
    ModuleData:
      type: object
      description: |-
        A struct containing data for multiple signaling modules, each associated
        with the module's namespace.
      additionalProperties: {}
      propertyNames:
        anyOf:
          - type: string
            description: A module identifier
            examples:
              - mymodule
            maxLength: 255
            minLength: 1
            pattern: "^[_0-9a-z]*$"
        default: core
        example: mymodule
      example:
        livekit:
          api_key: devkey
          api_secret: secret
          public_url: "http://localhost:7880"
          service_url: "http://localhost:7880"
...
    PublicUserProfile:
      allOf:
        - $ref: "#/components/schemas/UserInfo"
          description: General information about the user
        - type: object
          required:
            - id
            - email
            - timezone
          properties:
            email:
              type: string
              description: The email of the user
            id:
              $ref: "#/components/schemas/UserId"
              description: The user id
            timezone:
              $ref: "#/components/schemas/TimeZone"
              description: The local timezone of the user
      description: |-
        Public user details.

        Contains general "public" information about a user. Is accessible to all other users.
      example:
        avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
        display_name: Alice Adams
        email: alice@example.com
        firstname: Alice
        id: 00000000-0000-0000-0000-0000000a11c3
        lastname: Adams
        timezone: Europe/Berlin
        title: ""
...
    RoomId:
      type: string
      format: uuid
      description: The id of a room
      example: 00000000-0000-0000-0000-0000abadcafe
    RoomParameters:
      type: object
      required:
        - created_by
        - waiting_room
        - tariff
        - streaming_links
        - e2e_encryption
      properties:
        call_in:
          $ref: "#/components/schemas/CallInInfo"
        created_by:
          $ref: "#/components/schemas/PublicUserProfile"
        e2e_encryption:
          type: boolean
          description: Indicates whether the meeting room should have e2e encryption enabled.
        event:
          $ref: "#/components/schemas/EventContext"
        invite_code:
          $ref: "#/components/schemas/InviteCode"
        module_data:
          $ref: "#/components/schemas/ModuleData"
          description: Additional configuration options that are used by modules during initialization.
        password:
          $ref: "#/components/schemas/RoomPassword"
        streaming_links:
          type: array
          items:
            $ref: "#/components/schemas/StreamingLink"
        tariff:
          $ref: "#/components/schemas/TariffResource"
        waiting_room:
          type: boolean
      example:
        call_in:
          id: "1234567890"
          password: "0987654321"
          tel: +555-123-456-789
        created_by:
          avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
          display_name: Alice Adams
          email: alice@example.com
          firstname: Alice
          id: 00000000-0000-0000-0000-0000000a11c3
          lastname: Adams
          timezone: Europe/Berlin
          title: ""
        e2e_encryption: false
        event:
          description: The Weekly Team Event
          ends_at: "1970-01-01T01:00:00Z"
          id: 00000000-0000-0000-0000-004433221100
          is_adhoc: false
          shared_folder:
            read:
              password: v3rys3cr3t
              url: "https://cloud.example.com/shares/abc123"
          starts_at: "1970-01-01T00:00:00Z"
          title: Team Event
        invite_code: 00000000-0000-0000-0000-0000deadbeef
        module_data:
          livekit:
            api_key: devkey
            api_secret: secret
            public_url: "http://localhost:7880"
            service_url: "http://localhost:7880"
        password: "1234"
        streaming_links:
          - name: My OwnCast Stream
            url: "https://owncast.example.com/mystream"
        tariff:
          id: 00000000-0000-0000-0000-000000000000
          modules:
            chat:
              features: []
            core:
              features: []
            livekit:
              features: []
            moderation:
              features: []
            recording:
              features:
                - record
          name: Starter tariff
          quotas:
            max_storage: 50000
        waiting_room: false
    RoomPassword:
      type: string
      description: A room password
      examples:
        - v3rys3cr3t
      maxLength: 255
      minLength: 1
...
    StreamingLink:
      type: object
      description: Streaming link for an event
      required:
        - name
        - url
      properties:
        name:
          type: string
          description: The name of the streaming link
        url:
          type: string
          format: uri
          description: The url of the streaming link
      example:
        name: My OwnCast Stream
        url: "https://owncast.example.com/mystream"
...
    TariffResource:
      type: object
      description: Information related to a specific tariff
      required:
        - id
        - name
        - quotas
        - modules
      properties:
        id:
          $ref: "#/components/schemas/TariffId"
          description: The ID of the tariff
        modules:
          type: object
          description: "Enabled modules for the tariff, including their enabled features"
          additionalProperties:
            $ref: "#/components/schemas/TariffModuleResource"
          propertyNames:
            anyOf:
              - type: string
                description: A module identifier
                examples:
                  - mymodule
                maxLength: 255
                minLength: 1
                pattern: "^[_0-9a-z]*$"
            default: core
            example: mymodule
        name:
          type: string
          description: The name of the tariff
        quotas:
          type: object
          description: The quotas of the tariff
          additionalProperties:
            type: integer
            format: int64
            minimum: 0
          propertyNames:
            anyOf:
              - type: string
                description: A quota type
                examples:
                  - max_storage
            example: max_storage
      example:
        id: 00000000-0000-0000-0000-000000000000
        modules:
          chat:
            features: []
          core:
            features: []
          livekit:
            features: []
          moderation:
            features: []
          recording:
            features:
              - record
        name: Starter tariff
        quotas:
          max_storage: 50000
    TimeZone:
      type: string
      description: A time zone
      examples:
        - Europe/Berlin
    Token:
      type: string
      format: uuid
      description: The token that is required to start a roomserver signaling session
      example: 00000000-0000-0000-0000-0000abadcafe
    TokenRequestBody:
      type: object
      description: "The request body for GET `/rooms/{room_id}/token`"
      required:
        - client_parameters
      properties:
        client_parameters:
          $ref: "#/components/schemas/ClientParameters"
          description: Information regarding the requesting client
        room_parameters:
          $ref: "#/components/schemas/RoomParameters"
          description: |-
            The room that the token is requested for.

            Once this was provided to the roomserver, further request do not need to include the room parameters
      example:
        client_parameters:
          device_secret: v3rys3cr3tD3v1ce5tr1ng
          kind: registered
          profile:
            avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
            display_name: Alice Adams
            email: alice@example.com
            firstname: Alice
            id: 00000000-0000-0000-0000-0000000a11c3
            lastname: Adams
            timezone: Europe/Berlin
            title: ""
          role: moderator
        room_parameters:
          call_in:
            id: "1234567890"
            password: "0987654321"
            tel: +555-123-456-789
          created_by:
            avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
            display_name: Alice Adams
            email: alice@example.com
            firstname: Alice
            id: 00000000-0000-0000-0000-0000000a11c3
            lastname: Adams
            timezone: Europe/Berlin
            title: ""
          e2e_encryption: false
          event:
            description: The Weekly Team Event
            ends_at: "1970-01-01T01:00:00Z"
            id: 00000000-0000-0000-0000-004433221100
            is_adhoc: false
            shared_folder:
              read:
                password: v3rys3cr3t
                url: "https://cloud.example.com/shares/abc123"
            starts_at: "1970-01-01T00:00:00Z"
            title: Team Event
          invite_code: 00000000-0000-0000-0000-0000deadbeef
          module_data:
            livekit:
              api_key: devkey
              api_secret: secret
              public_url: "http://localhost:7880"
              service_url: "http://localhost:7880"
          password: "1234"
          streaming_links:
            - name: My OwnCast Stream
              url: "https://owncast.example.com/mystream"
          tariff:
            id: 00000000-0000-0000-0000-000000000000
            modules:
              chat:
                features: []
              core:
                features: []
              livekit:
                features: []
              moderation:
                features: []
              recording:
                features:
                  - record
            name: Starter tariff
            quotas:
              max_storage: 50000
          waiting_room: false
    TokenResponse:
      oneOf:
        - type: object
          description: The signaling token for the requested room
          required:
            - token
            - response
          properties:
            response:
              type: string
              enum:
                - token
            token:
              $ref: "#/components/schemas/Token"
        - type: object
          description: "The requested room is unknown to the roomserver and needs to be provided through the [TokenRequestBody](super::TokenRequestBody)"
          required:
            - response
          properties:
            response:
              type: string
              enum:
                - unknown_room
      description: "The response body for GET `/rooms/{room_id}/token`"
      example:
        response: token
        token: 00000000-0000-0000-0000-0000abadcafe
    UserId:
      type: string
      format: uuid
      description: The id of a user
      example: 00000009-9889-9889-9889-988000000000
    UserInfo:
      type: object
      description: Information about the room creator
      required:
        - title
        - firstname
        - lastname
        - display_name
        - avatar_url
      properties:
        avatar_url:
          type: string
          description: The creators avatar url
        display_name:
          $ref: "#/components/schemas/DisplayName"
          description: The creators display name
        firstname:
          type: string
          description: The creators first name
        lastname:
          type: string
          description: The creators last name
        title:
          $ref: "#/components/schemas/UserTitle"
          description: Optional title of the creator
      example:
        avatar_url: "https://gravatar.com/avatar/c160f8cc69a4f0bf2b0362752353d060"
        display_name: Alice Adams
        firstname: Alice
        lastname: Adams
        title: ""
...
  securitySchemes:
    API-Token:
      type: http
      scheme: bearer
      bearerFormat: api token
      description: "The roomservers API token is expected to be in the `Authorization` header with the format: `bearer <token>`"
...